backend-functions/requirements.txt
azure-core==1.40.0
azure-functions==2.1.0
azure-storage-blob==12.28.0
certifi==2026.4.22
cffi==2.0.0
charset-normalizer==3.4.7
cryptography==47.0.0
et_xmlfile==2.0.0
idna==3.13
isodate==0.7.2
MarkupSafe==3.0.3
numpy==2.4.4
openpyxl==3.1.5
pandas==3.0.2
pycparser==3.0
pyodbc==5.3.0
python-dateutil==2.9.0.post0
requests==2.33.1
six==1.17.0
typing_extensions==4.15.0
tzdata==2026.2
urllib3==2.6.3
Werkzeug==3.1.8
